server/requirements.txt
grpcio
h5py
matplotlib
numpy
opcua
pandas
protobuf
scikit_learn
scipy
tslearn
cryptography
